HTBHF-2544 refactor state machine selectors
