+++ src/web/routes/application/flow-control/state-machine/selectors/index.js
@@ -1,0 +1,6 @@
+const { getNextForStep } = require('./get-next-for-step')
+
+module.exports = {
+  ...require('./selectors'),
+  getNextForStep
+}
